feat: Fixes #75. Add edit icon from semantic-ui to account name input field (#83)
* feat: Fixes #75. Add edit icon from semantic-ui to account name input field

Adding an edit icon that is embedded in the account name input field makes it more obvious to the user that the input field isn't read-only, and that there is a way to "click before doing" (i.e. editing the input field that is identified as being editable before the ability to save changes is granted). Some users may click the edit icon itself thinking it toggles the ability to edit the input field, whilst others may just recognise that the edit icon means the input field is editable. In either case just having the edit icon should suffice as being the initial "call to action" for a user wishing to edit the input field, and since the Reset and Save buttons illuminate when the account name is modified, they'll recognise that a further step is required to save their changes.

* resolve introduced ESLint and Flow errors

* fix: Move attributes up so they are in alphabetical order

* fix: Remove unnecessary classenames. Edit icon stays within input field